Use correct var type to fix tpm attestation report

+++ vim25/types/types.go
@@ -22116,7 +22116,7 @@ type HostDigestInfo struct {
 	DynamicData
 
 	DigestMethod string `xml:"digestMethod"`
-	DigestValue  []byte `xml:"digestValue"`
+	DigestValue  []int `xml:"digestValue"`
 	ObjectName   string `xml:"objectName,omitempty"`
 }
 
@@ -26446,7 +26446,7 @@ func init() {
 type HostTpmEventDetails struct {
 	DynamicData
 
-	DataHash       []byte `xml:"dataHash"`
+	DataHash       []int `xml:"dataHash"`
 	DataHashMethod string `xml:"dataHashMethod,omitempty"`
 }
